Escrowflow/requirements.txt
flask==2.3.2
python-telegram-bot==20.3
python-dotenv==1.0.0
